Fix start script

+++ start.sh
@@ -1,9 +1,14 @@
+set -e
+
 if [ -z "$DATABASE_PATH" ]; then
   echo "DATABASE_PATH is not set"
   exit 1
 fi
 
-cp "$DATABASE_PATH" "$DATABASE_PATH".bak"$(date +%s)"
+# If database doesn't exist, create it
+if [ -f "$DATABASE_PATH" ]; then
+  cp "$DATABASE_PATH" "$DATABASE_PATH".bak"$(date +%s)"
+fi
 
 export DATABASE_URL="sqlite:$DATABASE_PATH"
 bun x dbmate up
